# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Standout Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scenarios:
    get:
      operationId: listScenarios
      tags: [scenarios]
      summary: List scenarios
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
  /scenarios/{id}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get scenario by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scenario detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /scenarios/{id}/complete:
    post:
      operationId: completeScenario
      tags: [scenarios]
      summary: Complete a scenario with chosen response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteScenarioBody"
      responses:
        "200":
          description: Scenario completion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScenarioResult"
  /systems:
    get:
      operationId: listSystems
      tags: [systems]
      summary: List management systems/playbooks
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/System"
  /systems/{id}:
    get:
      operationId: getSystem
      tags: [systems]
      summary: Get system by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: System detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
...
  /admin/scenarios:
    get:
      operationId: adminListScenarios
      tags: [admin]
      summary: Admin list all scenarios
      responses:
        "200":
          description: All scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
    post:
      operationId: adminCreateScenario
      tags: [admin]
      summary: Admin create scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioBody"
      responses:
        "201":
          description: Created scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /admin/scenarios/{id}:
    put:
      operationId: adminUpdateScenario
      tags: [admin]
      summary: Admin update scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioBody"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
    delete:
      operationId: adminDeleteScenario
      tags: [admin]
      summary: Admin delete scenario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/systems:
    get:
      operationId: adminListSystems
      tags: [admin]
      summary: Admin list all systems
      responses:
        "200":
          description: All systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/System"
    post:
      operationId: adminCreateSystem
      tags: [admin]
      summary: Admin create system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemBody"
      responses:
        "201":
          description: Created system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Scenario:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        situation:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedMinutes:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioOption"
        tags:
          type: array
          items:
            type: string
        isCompleted:
          type: boolean
        userScore:
          type: integer
          nullable: true
        isPremium:
          type: boolean
      required: [id, title, description, situation, category, difficulty, estimatedMinutes, options, tags, isCompleted, isPremium]
    ScenarioOption:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        quality:
          type: string
          enum: [strong, acceptable, weak, risky]
        explanation:
          type: string
        brandImpact:
          type: string
        score:
          type: integer
      required: [id, text, quality, explanation, brandImpact, score]
    CompleteScenarioBody:
      type: object
      properties:
        selectedOptionId:
          type: string
      required: [selectedOptionId]
    ScenarioResult:
      type: object
      properties:
        scenarioId:
          type: string
        selectedOption:
          $ref: "#/components/schemas/ScenarioOption"
        score:
          type: integer
        feedback:
          type: string
        keyLesson:
          type: string
        brandImpactSummary:
          type: string
      required: [scenarioId, selectedOption, score, feedback, keyLesson, brandImpactSummary]
    ScenarioBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        situation:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedMinutes:
          type: integer
        options:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioOption"
        tags:
          type: array
          items:
            type: string
        isPremium:
          type: boolean
      required: [title, description, situation, category, difficulty, estimatedMinutes, options, tags, isPremium]
    System:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SystemStep"
        keyPrinciples:
          type: array
          items:
            type: string
        scripts:
          type: array
          items:
            $ref: "#/components/schemas/SystemScript"
        estimatedMinutes:
          type: integer
        isPremium:
          type: boolean
      required: [id, title, description, category, steps, keyPrinciples, scripts, estimatedMinutes, isPremium]
    SystemStep:
      type: object
      properties:
        order:
          type: integer
        title:
          type: string
        description:
          type: string
      required: [order, title, description]
    SystemScript:
      type: object
      properties:
        tone:
          type: string
        text:
          type: string
      required: [tone, text]
    SystemBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SystemStep"
        keyPrinciples:
          type: array
          items:
            type: string
        scripts:
          type: array
          items:
            $ref: "#/components/schemas/SystemScript"
        estimatedMinutes:
          type: integer
        isPremium:
          type: boolean
      required: [title, description, category, steps, keyPrinciples, scripts, estimatedMinutes, isPremium]
...