# lib/api-zod/src/generated/types/completeScenarioBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Standout Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CompleteScenarioBody {
  selectedOptionId: string;
}